[Pipes] for String parsing
Migrated Pipes away from App.module

pipe for parsing any string length into 2-char size array done
