+++ CharacterGenerator/src/app/app.module.ts
@@ -12,36 +12,7 @@ import { LoginComponent } from './components/login/login.component';
 import { ProfileComponent } from './components/profile/profile.component';
 import { InfoComponent } from './components/info/info.component';
 import { SpellsComponent } from './components/spells/spells.component';
-import { Multiclass } from './models/multiclass';
-
-@Pipe({name: 'round'})
-export class RoundPipe implements PipeTransform {
-    transform(value: number): number {
-        return Math.floor(value);
-    }
-}
-
-@Pipe({name: 'class'})
-export class ClassPipe implements PipeTransform {
-    transform(value: Multiclass[]): string {
-      if(value.length>1){return "Multiclass";}
-      else {
-        const classitem=value[0];
-        return classitem.charClass.classname;
-      }
-  }
-}
-@Pipe({name: 'level'})
-export class LevelPipe implements PipeTransform {
-    transform(value: Multiclass[]): number {
-      var totalLevel = 0;
-      for(let multiclass of value){
-        totalLevel+=multiclass.level;
-      }
-      return totalLevel;
-      }
-  }
-
+import { RoundPipe, ClassPipe, LevelPipe,StringParserPipe} from './pipes/custom.pipe';
 
 
 
@@ -55,9 +26,8 @@ export class LevelPipe implements PipeTransform {
     ProfileComponent,
     InfoComponent,
     SpellsComponent,
-    RoundPipe,
-    ClassPipe,
-    LevelPipe
+    RoundPipe, ClassPipe, LevelPipe,StringParserPipe,
+
   ],
   imports: [
     BrowserModule,
